Refactor: rename "headers" to parsing hints
